[ADD] integration methods of OpenAI,
