+++ src/users/usersController.ts
@@ -2,7 +2,8 @@ import { Request, Response } from "express";
 import { IRequest } from "src/interface";
 import prisma from "../../config/prisma";
 import { GENERAL_ERROR } from "../constants";
-import { isExistingUser } from "../utils";
+import { cryptoEncode, isExistingUser } from "../utils";
+import { sign } from "jsonwebtoken";
 
 export async function getUser(req: Request, res: Response) {
   try {
@@ -21,10 +22,18 @@ export async function postUsers(req: IRequest, res: Response) {
     if (await isExistingUser(req.body.email)) {
       throw new Error(GENERAL_ERROR);
     }
-    const response = await prisma.user.create({
+    req.body.password = await cryptoEncode(req.body.password);
+    const user = await prisma.user.create({
       data: req.body,
     });
-    res.status(200).send({ users: response });
+    // Implement Jwt Authentication
+    const result = await sign({
+      sub: {
+        userId: user?.id
+      },
+    }, process.env.SECRET!, { expiresIn: '3600' });
+    
+    res.status(200).send({ token: result });
   } catch (error: any) {
     console.log(error)
     res.status(400).send({ error: error.message ? error.message : GENERAL_ERROR });
